remove implicit dependencies on jQuery
For apps that explicitly remove their dependency on jQuery, this addon will break because it
imports from jquery assuming the consuming app already has it.

jQuery is imported in the component, but no dependency is declared in
the package.json. This works because until recently, jQuery was
available packaged with Ember source.

[Ember RFC 0386](https://emberjs.github.io/rfcs/0386-remove-jquery.html)
propses to remove jQuery eventually, and Ember already ships with the
@ember/jquery package installed directly as part of the "ember new"
output for new ember apps.

related to https://github.com/oskarrough/ember-youtube/issues/54